requirements.txt
asgiref==3.7.2
Django==4.2.2
django-filter==23.2
django-multiselectfield==0.1.12
django-nested-admin==4.0.2
djangorestframework==3.14.0
Pillow==9.5.0
python-monkey-business==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
